[MOV] test_auth_custom: merged into test_http
