--- odoo/addons/test_auth_custom/__manifest__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-{
-    'name': 'Tests that custom auth works & is not impaired by CORS',
-    'category': 'Hidden',
-    'author': 'Odoo S.A.',
-    'license': 'LGPL-3',
-}
