--- odoo/addons/test_auth_custom/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_endpoints
