+++ odoo/addons/test_http/tests/__init__.py
@@ -1,3 +1,4 @@
+from . import test_auth_custom
 from . import test_captcha
 from . import test_common
 from . import test_device
